Day 5, Part 2: Try benchmarking
I ended up moving all the code out of main into lib mostly so I could
easily benchmark things, and it was getting long enough that it felt
like it cluttered up main a bit.

I wanted to get some experience with criterion and benchmarking in
general, so I tested out a hypothesis about a couple of different
implementations of moving multiple items from one vec to another. I'm
not entirely sure I set up the benchmark correctly, but the results seem
to make sense to me. That the option that does not require an extra Vec
allocation every time is faster.