Day 16, Part 2: Refactor and tweak broken heuristic for part 1
In the previous commit, main.rs was just getting too big and cluttered
and hard to navigate. I split things up into a few different modules,
and introduced a Volcano struct which holds the tunnel/valve setup in
its state so it does't have to be explicitly passed around so much.
Other than that I discovered that my optimization heuristic for part 2
actually broke part 1, so I had to apply different heuristics for each.
This is a clear indication that my code won't work for any arbitrary
input, and so I should find a better solution. Plus it's still crazy
slow and takes a ton of memory.